pyproject.toml
[tool.poetry]
name = "paper-rag"
version = "0.1.0"
description = "Experimenting with RAG approaches for ArXiv and Wikipedia datasets"
authors = ["alexhalsey <alexander.halsey@utexas.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fire = "^0.7.0"
loguru = "^0.7.2"
llama-index-llms-openai = "^0.2.16"
pandas = "^2.2.3"
fastembed = "^0.4.1"
qdrant-client = "^1.12.0"
llama-index-vector-stores-qdrant = "^0.3.2"
llama-index-readers-file = "^0.2.2"
llama-index-embeddings-fastembed = "^0.2.0"
datasets = "^3.0.2"
llama-index-postprocessor-cohere-rerank = "^0.2.1"
streamlit = "^1.41.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
